third_party/WebKit/LayoutTests/external/wpt/orientation-event/devicemotionevent-init-expected.txt

This is a testharness.js-based test.
FAIL User created event is initialized properly with default values assert_equals: interval must be initialized to 0 expected (number) 0 but got (object) null
PASS User created event is initialized properly using dictionary 
Harness: the test ran to completion.
